fix(react): Add `POP` guard for long-running `pageload` spans (#17867)
This resolves the issue that occurs when an extra `navigation`
transaction is created after a prematurely ended `pageload` transaction
in React Router lazy routes.

This apparently occurs when there's a long-running pageload with
lazy-routes (after fetching assets, there are multiple potentially
long-running API calls happening).

This causes the `pageload` transaction to prematurely end, even before
the fully parameterized transaction name is resolved. The reason is that
there can be a `POP` event emitted, which we subscribe to create a
`navigation` transaction. This ends the ongoing `pageload` transaction
before its name is updated with a resolved parameterized route path, and
starts a `navigation` transaction, which contains the remaining spans
that were supposed to be a part of the `pageload` transaction.

This fix makes sure the initial `POP` events are not necessarily treated
as `navigation` pointers, which should fix both:

- Duplicate / extra `navigation` transactions having a part of
`pageload` spans.
- Remaining wildcards in the `pageload` transaction names