feat(firebase): Instrument cloud functions for firebase v2 (#17952)
Closes: #17861 

This adds
- instrumentation of Cloud Functions for Firebase (v2) along side the
Firestore integration. It can be used with the
`Sentry.firebaseIntegration()` (this is atm not documented in the docs
and got added in #16719, but will be added right after this has been
merged. See https://github.com/getsentry/sentry-docs/issues/15247).
- The test app for Firebase has been rewritten and updated since it
requires a little special setup.

<details>
<summary>Supported functions</summary>
<ul>
  <li>onRequest</li>
  <li>onCall</li>
  <li>onDocumentCreated</li>
  <li>onDocumentUpdated</li>
  <li>onDocumentDeleted</li>
  <li>onDocumentWritten</li>
  <li>onDocumentCreatedWithAuthContext</li>
  <li>onDocumentUpdatedWithAuthContext</li>
  <li>onDocumentDeletedWithAuthContext</li>
  <li>onDocumentWrittenWithAuthContext</li>
  <li>onSchedule</li>
  <li>onObjectFinalized</li>
  <li>onObjectArchived</li>
  <li>onObjectDeleted</li>
  <li>onObjectMetadataUpdated</li>
</ul>
</details>

Bear in mind that the OTel attributes for FaaS are still in
[Development](https://opentelemetry.io/docs/specs/semconv/faas/faas-spans/)
and could change or be removed over time (not sure if we should then
even add them in here at this point in time).