fix(react): Don't trim index route `/` when getting pathname (#17985)
Currently, the test in this PR fail:
https://github.com/getsentry/sentry-javascript/pull/17789

Root routes can yield an empty transaction name, causing `<unlabeled
transaction>` instead of `/` as the transaction name for the root.
This happens when the router includes children routes with `index:
true`.

The route matching is also depending on the `allRoutes` Set. The
`allRoutes` Set include the children routes twice (once as children of
the route and once as a root element of the Set). When only including
them once, it works but parametrization does not work anymore.
--> First I thought, this duplication is the cause but probably it isn't

## What’s broken
Root cause is in `sendIndexPath(...)`:
- Mis-parenthesized ternary picks `stripBasenameFromPathname` instead of
`pathBuilder`.
  - Trimming turns `/` into an empty string.