feat: Add a note to save changes before starting (#17987)
During the release I figured that most of the times you might work on
features/bugfixes. In that case the current release process this
scenario isn't specified and who knows what it might or might not do.

I added point 1. and 9. where we ask Cursor to save the changes (1) and
go back to the state it was before (9). This would make the release
process more complete.

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>