fix(nuxt): check for H3 error cause before re-capturing (#18035)
The flakey tests in nuxt-4 pointed out that we have a race condition
where a middleware error can bubble up as an H3 event error, which wraps
the original error we caught.

This means the
[`checkOrSetAlreadyCaught`](https://github.com/getsentry/sentry-javascript/blob/749638766641b552e1193353c3f0430cf970787d/packages/core/src/utils/misc.ts#L212-L232)
won't actually detect it since it doesn't check the [`.cause`
property](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/cause).

I added the logic needed for that for the Nuxt SDK but feels like this
may come up later if it hadn't already.

Note that this does not affect the spans created, just the mechanism of
the caught error, the spans would still be marked correctly as errored.