feat(core): Send `user-agent` header with envelope requests in server SDKs (#17929)
As a follow-up of an incident, SDKs were asked to send a `user-agent` Http header to
determine by just looking at incoming request headers which SDK the
request is coming from. See [develop
specification](https://develop.sentry.dev/sdk/overview/#user-agent).

This PR makes the following changes to sending user agent HTTP headers
with envelope requests made by the transport:

- Send `user-agent` in all server-runtime SDKs
- Extract the `headers` option from individual transport options to
`BaseTransportOptions`. This allows us to type-safely add the user agent
header to the SDKs transport options which is the easiest way to pass a
header to the transport without having to modify or extend any of the
existing transport APIs. I checked and every transport implementation we
currently export exposed a `headers` option anyway, so this just unifies
it. Given this is an optional property, custom transport implementations
extending `BaseTransportOptions` won't break either. The problem here is
only that they might not actually support this option which all things
considered I think is fine. If reviewers have different opinions, I'm
happy to revisit this.
- Unit and integration/e2e tests for node and cloudflare