fix(replay): Linked errors not resetting session id (#17854)
This PR fixes a case where we [correctly] tag an error event w/ replay
id, but something occurs where the replay event does not end up being
flushed. This means the existing session is still in a buffered state,
and will keep its session id until a new error event is sampled and a
replay is created. When this does happen, we can have a replay with a
super long duration (e.g. the time between the two error replays).

We now update the session immediately when we tag an error event w/
replay id so that if the replay event does not successfully flush, the
session will respect its expiration date.