fix(nextjs): Stop accessing sync props in template (#18113)
We need to stop accessing params within our wrapper templates as this
can break statically rendered pages and cause cache busts in dynamic
pages.

Opened https://github.com/getsentry/sentry-javascript/issues/18115 to
investigate other ways of handling the functionality we're removing
here.

closes https://github.com/getsentry/sentry-javascript/issues/18097
closes https://github.com/getsentry/sentry-javascript/issues/17431

Note: There's some unused code in the utils that I did not delete as we
might be able to reuse this in
https://github.com/getsentry/sentry-javascript/issues/18115