test(next): fix flakey tests (#18100)
I can't reproduce the flakiness locally, I logged a few things on CI
runs and it seems while the fetch spans does get created, it doesn't get
created on the middletrasnaction.

This is similar to what the note is saying in dev-mode, for it to be
less flakey I'm asserting if it either has the span on the right
transaction or if it got created in a separate transaction. Ideally we
want it to be created within the middleware transaction but it seems
like there is a timing issue preventing it from being picked up there.

At any case this seems to reduce or eliminate the chances of failed
runs, previously it would fail consistently but with this it didn't fail
in 5 runs so far.