feat(aws): Add `SENTRY_LAYER_EXTENSION` to configure using the lambda layer extension via env variables (#18101)
Our recommended approach for the aws sdk is to use our Lambda layer.
When the Lambda layer is used, it would previously automatically set up
the layer extension. This can be disabled by setting `useLayerExtension:
false` during init, but when using the layer with our auto-wrapping as
per recommendation, it wasn't possible to opt out of using the
extension.

This PR adds `SENTRY_LAYER_EXTENSION` that can be used to opt out of the
extension.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>