fix(core): Only consider exception mechanism when updating session status from event with exceptions (#18137)
This PR fixes an oversight in our session update logic when we update
the SDK's session status while processing an error event. Previously,
we'd always set `status: 'crashed'` as soon as the top-level
`event.level` field was set to `'fatal'` (added in
https://github.com/getsentry/sentry-javascript/pull/15072). However,
this disregarded the case where the level was initially `'fatal'` but
later on the SDK or users decided to set the
`event.exceptions.values[i].mechanism.handled` field to `true`.

Simplest reproduction:

```
Sentry.captureException(new Error('test'), { captureContext: { level: 'fatal' } });
```

This PR changes the update logic to only look at the event mechanisms if
the event contained exceptions. I _think_ this is still compatible with
#15072 because the PR suggests we only care about `event.level ===
'fatal'` if no exceptions are on the event. But I'd appreciate a second
pair of eyes on this (cc @timfish).