feat(core): Truncate request string inputs in OpenAI integration (#18136)
The main OpenAI APIs are the
[completions](https://platform.openai.com/docs/api-reference/completions/create)
and
[responses](https://platform.openai.com/docs/api-reference/responses/create)
API. Currently we truncate messages sent to the API only if the user
input is an array of strings. This works for the completions, but not
for the responses API (where the input is a plain string).

Updated the truncation logic to also truncate if the input is a plain
string (+ test to verify everything works now as expected).