fix(core): Flatten gen_ai.request.available_tools in google-genai (#18194)
While investigating [this
ticket](https://linear.app/getsentry/issue/JS-657/available-tools-json-should-be-a-stringified-json-array-of-objects-not)
I noticed that available tools are sent as a nested instead of a flat
array in google genai, which seems like a bug to me.

The format I would expect and how we do it in other integrations is:
[{tool-definition}, {tool-definition}]

What we actually send atm is:
[[{tool-definition}], [{tool-definition}]]

This PR fixes this to instead send a flat list of tool definitions.