fix(core): Stringify available tools sent from vercelai (#18197)
[Linear
Ticket](https://linear.app/getsentry/issue/JS-657/available-tools-json-should-be-a-stringified-json-array-of-objects-not)

The available tools sent from our SDKs should generally be in the format
of a stringified array of objects (where an object stores information
about a particular tool). This is true for all AI SDKs except Vercel,
where we send an array of strings. This PR fixes this by parsing the
available tool array and converting the whole array into a proper string
representation.