fix(nextjs): Drop meta trace tags if rendered page is ISR (#18192)
**Summary**

ISR pages will have a `sentry-trace` and `baggage` meta tags rendered on
them following the initial render or after the first invalidation
causing a cached trace id to be present until the next invalidation.

This happens in Next.js 15/16 and both on Turbopack and Webpack.


**What I tried and didn't work**

I Found no way to conditionally set/unset/change the values set by the
`clientTraceMetadata` option, I found nothing useful on unit async
storages, nor re-setting the propagation context works. The
`clientTraceMetadata` gets called way earlier at the `app-render.tsx`
level, which would call our `SentryPropagator.inject()` then. We cannot
intercept it either because it runs before the page wrapper is called.

The main issue is _timing_:

- Suppressing the tracing wouldn't work either because it is too late.
Ideally we want a way to tell Next to remove those attributes at
runtime, or render them conditionally.
- I also tried setting everything that has to do with `sentry-trace` or
baggage to dummy values as some sort of "marker" for the SDK on the
browser side to drop them, but again it is too late since
`clientTraceMetadata` is picked up too early.


**Implementation**

so I figured a workaround, I decided to do it on the client side by:

- Marking ISR page routes via the route manifest we already have.
- In `Sentry.init` call we remove the tags before the browser
integration has had a chance to grab the meta tags.

Not the cleanest way, but I verified the issue by writing tests for it
and observing page loads across multiple page visits having the same
trace id. The meta deletion forces them to have new id for every visit
which is what we want.