feat(metrics): Add default `server.address` attribute on server runtimes (#18242)
Attaches a `server.address` attribute to all captured metrics on a
`serverRuntimeClient`

Did this by emitting a new `processMetric` hook in core, that we listen
to in the `serverRuntimeClient`. This way we do not need to re-export
all metrics functions from server runtime packages and still only get a
minimal client bundle size bump.

Added integration tests for node + cloudflare

closes https://github.com/getsentry/sentry-javascript/issues/18240
closes
https://linear.app/getsentry/issue/JS-1178/attach-serveraddress-as-a-default-attribute-to-metrics

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>