chore: Add `bump_otel_instrumentations` cursor command (#18253)
Bumping OpenTelemetry instrumentations is an important but tedious task,
all instrumentations have to be bumped in lockstep across the codebase.
That includes easy to miss dev-packages and third party instrumentations
like prisma's.

This command should make it easier to do that.

Example of a PR that was kicked off with this command:
https://github.com/getsentry/sentry-javascript/pull/18239