ref(react): Add more guarding against wildcards in lazy route transactions (#18155)
Building on top of  #17962 

Added a few more checks to make sure non-resolved (wildcard) routes are
not reported in lazy route pageloads / navigations.

- Improved `patchSpanEnd` with a user-configurable wait timeout for
potentially slow route resolution. Named this option as
`lazyRouteTimeout` and it's defaulted as `idleTimeout` * 3. It may
conditionally delay reporting (if the route resolution is still not done
by the end of the timeout), but will prevent prematurely sent lazy-route
transactions inside that window.
- Added extra checks on `updateNavigationSpan` and `handleNavigation`
for whether any wildcard still exists in a lazy-route, so they are still
marked as open to full resolution. We keep track of pending lazy-route
resolutions inside `pendingLazyRouteLoads`
- Added a final attempt to update the transaction name with
fully-resolved route when the pending resolution is done.

Any of these should not affect the behaviour of non-lazy route usage

---------

Co-authored-by: Sigrid <32902192+s1gr1d@users.noreply.github.com>