feat(core): Re-add `_experiments.enableLogs` option (#18299)
We're re-introducing
`_experiments.enableLogs`. The option stays deprecated and maybe we can
actually remove it or type it as `undefined` in the next major to sunset
it for good. Main motivation for re-adding: The flag was introduced in
v9 while we already worked on v10 where we removed it again. Therefore,
it had an unusually short lifespan. Some users didn't realize this when
upgrading to v10 and were wondering where their logs went.