feat(nextjs): Support `cacheComponents` on turbopack (#18304)
With this PR we stop breaking builds by simply re-exporting APIs that
make use of synchronous random APIs:
- `startSpan`
- `startSpanManual`
- `startInactiveSpan`

We do this with a naive approach of detecting a prod build phase and
returning a noop span in cases where we call these APIs. We hope to get
a more solid API for this from Vercel at some point so we can make this
more bulletproof. The current version will also still emit a warning in
webpack as the behaviour seems to change there.

This will only work for Turbopack as of now, as we still need to get rid
of build time instrumentation in webpack in order for this to work.

closes https://github.com/getsentry/sentry-javascript/issues/17895