feat(nextjs): Add cloudflare `waitUntil` detection (#18336)
This is part of the work done for supporting Next.js on cloudflare
workers in #14931, one of the issues is our flusher didn't take
cloudflare into account and the events were being cut off due to early
shutdowns by the worker environment.

We cannot use our core's `flushIfServerless` because it requires
explicit `cloudflareWaitUntil` to be passed down to it.

Also The symbol we are grabbing here is [specific to
opennext](https://github.com/opennextjs/opennextjs-cloudflare/blob/b53a046bd5c30e94a42e36b67747cefbf7785f9a/packages/cloudflare/src/cli/templates/init.ts#L17)
and isn't something that is globally available on the worker
environment. I had initially created #18330 because I mistook it for a
worker environment API thing.

So this PR adds that detection to Next.js since it is relevant here and
will use it if available, local tests show flushing is done correctly
and events aren't being cut off.