fix(core): continueTrace doesn't propagate given trace ID if active span exists (#18328)
During investigation of [this
issue](https://github.com/getsentry/sentry-javascript/issues/18296), we
discovered that `continueTrace` does not propagate the given trace ID if
it is called from within an already active span (in this instance this
happened because of the already active span started by the cloudflare
auto-instrumentation.

Solution: This PR resolves this behavior by setting the currently active
span to undefined when forking the propagation context in
`continueTrace`.

Tests:
- After introducing the fix one of the `@sentry/sveltekit` unit tests
started to fail. Not sure why this was triggered by this change, but I
fixed it by setting the `isSubRequest` property on the event instead of
the route, which seems correct to me.
- Added a unit test in `@sentry/core` that calls `continueTrace` inside
an active span and checks that the trace id is now propagated as we
would expect.