chore(ci): Add action to track all PRs as issues
This workflow checks PR descriptions (like this one) to see if it can find an
associated GitHub or Linear issue ID. If nothing is found, the workflow will
create a new GitHub Issue which will then be referenced in this PR. As all other
GitHub issues, this newly created one will then be synced to Linear as well.

This will increase visibility of (community) PRs in external tools like Linear.

Based on: https://github.com/getsentry/sentry-cocoa/pull/6064 by
@stephanie-anderson