feat: update eslint to v8 (#18264)
This upgrades our internal eslint config to use v8. Theoretically this
is considered a breaking change. Since this is an internal
`@sentry-internal/*` package, we can release this in a non-breaking
fashion

With v8 couple of changes came. Some of them are affecting us, which
required file changes:

-
[simple-import-sort/import](https://www.npmjs.com/package/eslint-plugin-simple-import-sort)
got couple of fixes
- "Unused eslint-disable directive": Unused "eslint-disable" comments
are now detected and marked as error
- [import/export](https://www.npmjs.com/package/eslint-plugin-import)
got a false positive, which affects most of our index barrel files (I
didn't want to disable it entirely, so I added a global disable in each
affected barrel file)
- tracked here:
https://github.com/import-js/eslint-plugin-import/issues/703


## Merge checks

- [ ] Rebased before merge to get the latest changes updated with the
new version