feat(nextjs): Deprecate Webpack top-level options (#18343)
This PR deprecates the webpack-only configurations via the `@deprecated`
JSDoc annotation and introduces a new `webpack` config namespace for
them.

Under the hood the logic was changed to read from the new values, with a
compatibility layer that sets them from the deprecated top-level options
while warning for each option if used.

This should set us up for a v11/v12 deletion of those options.

I might have missed a few options that only affect webpack, so I
appreciate a good look at this. At any case this isn't breaking so even
if missed a few, users won't experience disruptions.