chore(bugbot): Add testing conventions code review rules (#18433)
This PR adds some rules to bugbot's rulest to flag some testing-related
issues we'd like to avoid. Like with all AI rules, there are for sure
exceptions to the new rules, so no problem with us ignoring any of these
flags. But I think having an additional reminder that testing is
necessary would be a good change. LMK what you think!