fix(browser): Stringify span context in linked traces log statement (#18376)
I came upon this log `Sentry Logger [log]: Adding previous_trace [object
Object] link to span [object Object]` and this PR fixes this by
stringifying the context.

One concern I have with that is that the object could be too large
(stringifying takes too long) or circular. But this should be very
unlikely in this case. However, if someone else shares this concerns we
might change the log to either limit the depth or to only log specific
entries of the object (might add bundle size).


Closes #18377