fix(google-cloud-serverless): Move @types/express to optional peerDeps (#18452)
(closes #18438)
(closes
[JS-1273](https://linear.app/getsentry/issue/JS-1273/bump-typesexpress-to-v5-for-sentrygoogle-cloud-serverless))

This moves the express types as optional peerDependencies. Since we are
only relying on `Request` and `Response` and nothing more, this should
be a save update. In order to also test against this - I updated the the
local types to v5.

The pattern of having the types as optional peerDependencies is already
given for our Cloudflare SDK:


https://github.com/getsentry/sentry-javascript/blob/2ef3938fecf872b3d09006538484e5de97123ac5/packages/cloudflare/package.json#L55-L62