feat(replay): Add Request body with `attachRawBodyFromRequest` option (#18501)
Previously, we only got the body from the `fetch` options. However, some
people use the `Request` object which can also contain a body. As this
body is a `ReadableStream`, we could only read that async.

With `attachRawBodyFromRequest`, the `Request` object is patched to
include the raw body as a `Symbol` so it can be read synchronously.

Linear ticket (internal):
https://linear.app/getsentry/issue/JS-1255/investigate-sdk-issues-causing-all-requests-to-default-to-get