fix(browser): Reduce number of `visibilitystate` and `pagehide` listeners (#18581)
Previously, we called `onHidden` in our slightly modified version of the
vendored in `whenIdleOrHidden` API from web-vitals. This caused 2
additional event listeners to be registered with each `whenIdleOrHidden`
invocation, which also didn't get cleaned up properly.

To fix this, and prevent similar situations, this PR:
- inlines the `pagehide` event listener registration in
`whenIdleOrHidden` so that we can remove the `onHidden` call
- adds `once: true` to the listener registration so that the callback is
only invoked once
- deprecates `onHidden` because IMHO we should remove it in v11 and
replace the one remaining use of it with a direct `visibilityChange`
subscription.

Closes #18584 (added automatically)

closes
https://linear.app/getsentry/issue/JS-1339/investigate-multiple-event-listeners-in-nextjs-sdk