chore(tests): Add unordered mode to cloudflare test runner (#18596)
So far if multiple expected envelopes were chained on the cloudflare
test runner, the runner expected them to arrive in the order pre-defined
by the caller, else the test would fail. This can sometimes be too
strict, resulting in flakes. This PR adds an unordered mode to the
runner that lifts the order restriction on incoming envelopes.

This mode is also used in the hono cloudflare test now.

Closes https://github.com/getsentry/sentry-javascript/issues/18589