fix(cloudflare): Consume body of fetch in the Cloudflare transport (#18545)
(closes #18534)
(closes
[JS-1319](https://linear.app/getsentry/issue/JS-1319/cloudflare-transport-doesnt-consume-fetch-response-bodies))

This consumes the body so it is safe to be closed by Cloudflare Workers.
Unfortunately this cannot be reproduced locally and is only happening
when the worker is being deployed (so there is no E2E test for this)

@AbhiPrasad do you think it is necessary to add this code snippet in the
other transports we have? It is what I've read specifically to
Cloudflare, since they seem to be super strict on it, but this could
potentially be a problem in other runtimes as well.