chore(react): Inline `hoist-non-react-statics` package (#18102)
Ref / Resolves:
https://github.com/getsentry/sentry-javascript/issues/18089

`hoist-non-react-statics` package we depend on is not ESM compatible,
and causing issues for users on Cloudflare environments. As this package
has not been updated in 5 years, and our use has not changed for long
time, I think we can just inline this to make it compatible, and save
users from explicitly externalizing this package.

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>