fix(node): relax Fastify's `setupFastifyErrorHandler` argument type (#18620)
When using
[exactOptionalPropertyTypes](https://www.typescriptlang.org/tsconfig/#exactOptionalPropertyTypes),
it is pretty hard to match Fastify's instance types given all the
missing overloads from the partial signature we have. The problematic
area is around `addHook`.

We either need to have all the signatures implemented exactly, or
outright import Fastify's types which is not ideal since it is not a
dependency.

I opted to relax the types specifically for `setupFastifyErrorHandler`
by providing a minimal instance with the one method it needs to avoid TS
trying to matching the other properties and methods signatures, not
ideal but simple enough.

I verified that this works for Fastify v3 throughout v5

closes #18619